Revert "GM9PRO_sprout: replace init.qcom.usb.sh to stock pie"
This reverts commit 1dbd5e4d3bf2e7cbe5b9090a46a00a705a832a89.

+++ rootdir/etc/init.qcom.usb.sh
@@ -29,60 +29,36 @@
 #
 
 # Set platform variables
-if [ -f /sys/devices/soc0/hw_platform ]; then
-    soc_hwplatform=`cat /sys/devices/soc0/hw_platform` 2> /dev/null
-else
-    soc_hwplatform=`cat /sys/devices/system/soc/soc0/hw_platform` 2> /dev/null
-fi
-
-if [ -f /sys/devices/soc0/machine ]; then
-    soc_machine=`cat /sys/devices/soc0/machine` 2> /dev/null
-else
-    soc_machine=`cat /sys/devices/system/soc/soc0/machine` 2> /dev/null
-fi
+soc_hwplatform=`cat /sys/devices/soc0/hw_platform 2> /dev/null`
+soc_machine=`cat /sys/devices/soc0/machine 2> /dev/null`
+soc_machine=${soc_machine:0:2}
+soc_id=`cat /sys/devices/soc0/soc_id 2> /dev/null`
 
 #
-# Check ESOC for external MDM
+# Check ESOC for external modem
 #
-# Note: currently only a single MDM is supported
+# Note: currently only a single MDM/SDX is supported
 #
-if [ -d /sys/bus/esoc/devices ]; then
-for f in /sys/bus/esoc/devices/*; do
-    if [ -d $f ]; then
-    if [ `grep -e "^MDM" -e "^SDX" $f/esoc_name` ]; then
-            esoc_link=`cat $f/esoc_link`
-            break
-        fi
-    fi
-done
-fi
+esoc_name=`cat /sys/bus/esoc/devices/esoc0/esoc_name 2> /dev/null`
 
 target=`getprop ro.board.platform`
+product=`getprop ro.product.name`
+product=${product:(-4)}
 
-# soc_ids for 8937
-if [ -f /sys/devices/soc0/soc_id ]; then
-	soc_id=`cat /sys/devices/soc0/soc_id`
-else
-	soc_id=`cat /sys/devices/system/soc/soc0/id`
+if [ -f /sys/class/android_usb/f_mass_storage/lun/nofua ]; then
+	echo 1  > /sys/class/android_usb/f_mass_storage/lun/nofua
 fi
 
 #
-# Allow USB enumeration with default PID/VID
+# Override USB default composition
 #
-baseband=`getprop ro.baseband`
-
-echo 1  > /sys/class/android_usb/f_mass_storage/lun/nofua
-usb_config=`getprop persist.vendor.usb.config`
-
-#TEL_L9501_A01-58/142/436 add by wangyibo at 20181011 start
-buildvariant1=`getprop ro.build.type`
-#TEL_L9501_A01-58/142/436 add by wangyibo at 20181011 end
-
-if [ "$usb_config" == "" ]; then #USB persist config not set, select default configuration
-      if [ "$esoc_link" != "" ]; then
+# If USB persist config not set, set default configuration
+if [ "$(getprop persist.vendor.usb.config)" == "" -a \
+	"$(getprop init.svc.vendor.usb-gadget-hal-1-0)" != "running" ]; then
+    if [ "$esoc_name" != "" ]; then
 	  setprop persist.vendor.usb.config diag,diag_mdm,qdss,qdss_mdm,serial_cdev,dpl,rmnet,adb
-      else
-	  case "$baseband" in
+    else
+	  case "$(getprop ro.baseband)" in
 	      "apq")
 	          setprop persist.vendor.usb.config diag,adb
 	      ;;
@@ -92,10 +68,13 @@ if [ "$usb_config" == "" ]; then #USB persist config not set, select default con
 	              setprop persist.vendor.usb.config diag,adb
 	          ;;
                   *)
-		  soc_machine=${soc_machine:0:3}
 		  case "$soc_machine" in
-		    "SDA")
-	              setprop persist.vendor.usb.config diag,adb
+		    "SA")
+			if [ "$product" == "gvmq" ]; then
+				setprop persist.vendor.usb.config adb
+			else
+				setprop persist.vendor.usb.config diag,adb
+			fi
 		    ;;
 		    *)
 	            case "$target" in
@@ -111,7 +90,6 @@ if [ "$usb_config" == "" ]; then #USB persist config not set, select default con
 			    else
 			               case "$soc_id" in
 				               "313" | "320")
-						  echo BAM2BAM_IPA > /sys/class/android_usb/android0/f_rndis_qc/rndis_transports
 				                  setprop persist.vendor.usb.config diag,serial_smd,rmnet_ipa,adb
 				               ;;
 				               *)
@@ -128,21 +106,12 @@ if [ "$usb_config" == "" ]; then #USB persist config not set, select default con
 			      fi
 		      ;;
 	              "msm8998" | "sdm660" | "apq8098_latv")
-#TEL_L9501_A01-58/142/436 add by wangyibo at 20181011 start
-				case "$buildvariant1" in
-					"userdebug" | "eng")
-		          		setprop persist.vendor.usb.config diag,serial_cdev,rmnet,adb
-					;;
-					"user")
-					setprop persist.sys.usb.config mass_storage
-					;;
-				esac
-#TEL_L9501_A01-58/142/436 add by wangyibo at 20181011 end
+		          setprop persist.vendor.usb.config diag,serial_cdev,rmnet,adb
 		      ;;
 	              "sdm845" | "sdm710")
 		          setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,adb
 		      ;;
-	              "msmnile")
+	              "msmnile" | "sm6150" | "trinket" | "lito" | "atoll" | "bengal")
 			  setprop persist.vendor.usb.config diag,serial_cdev,rmnet,dpl,qdss,adb
 		      ;;
 	              *)
@@ -158,16 +127,32 @@ if [ "$usb_config" == "" ]; then #USB persist config not set, select default con
       fi
 fi
 
-# set device mode notification to USB driver for SA8150 Auto ADP
-product=`getprop ro.build.product`
-
-case "$product" in
-	"msmnile_au")
-	echo peripheral > /sys/bus/platform/devices/a600000.ssusb/mode
-         ;;
-	*)
-	;;
+# Start peripheral mode on primary USB controllers for Automotive platforms
+case "$soc_machine" in
+    "SA")
+	if [ -f /sys/bus/platform/devices/a600000.ssusb/mode ]; then
+	    default_mode=`cat /sys/bus/platform/devices/a600000.ssusb/mode`
+	    case "$default_mode" in
+		"none")
+		    echo peripheral > /sys/bus/platform/devices/a600000.ssusb/mode
+		;;
+	    esac
+	fi
+    ;;
 esac
+
+# set rndis transport to BAM2BAM_IPA for 8920 and 8940
+if [ "$target" == "msm8937" ]; then
+	if [ ! -d /config/usb_gadget ]; then
+	   case "$soc_id" in
+		"313" | "320")
+		   echo BAM2BAM_IPA > /sys/class/android_usb/android0/f_rndis_qc/rndis_transports
+		;;
+		*)
+		;;
+	   esac
+	fi
+fi
 
 # check configfs is mounted or not
 if [ -d /config/usb_gadget ]; then
@@ -176,16 +161,15 @@ if [ -d /config/usb_gadget ]; then
 	msm_serial_hex=`printf %08X $msm_serial`
 	machine_type=`cat /sys/devices/soc0/machine`
 	product_string="$machine_type-$soc_hwplatform _SN:$msm_serial_hex"
-    # begin yutao modify for TEL_L9501_A01-60
-	echo "GM 9 Pro" > /config/usb_gadget/g1/strings/0x409/product
-    # end yutao modify for TEL_L9501_A01-60
+	echo "$product_string" > /config/usb_gadget/g1/strings/0x409/product
 
 	# ADB requires valid iSerialNumber; if ro.serialno is missing, use dummy
-	serialnumber=`cat /config/usb_gadget/g1/strings/0x409/serialnumber` 2> /dev/null
+	serialnumber=`cat /config/usb_gadget/g1/strings/0x409/serialnumber 2> /dev/null`
 	if [ "$serialnumber" == "" ]; then
 		serialno=1234567
 		echo $serialno > /config/usb_gadget/g1/strings/0x409/serialnumber
 	fi
+	setprop vendor.usb.configfs 1
 fi
 
 #
